Move safeupdate to utils
