+++ ftc/utils.py
@@ -1,3 +1,6 @@
+from copy import deepcopy
+from functools import reduce
+
 from ftc.registration import registry
 
 
@@ -13,3 +16,25 @@ def get_controllers(*args, env=None):
         controllers.append(make(id, env=env))
 
     return controllers or [Controller() for Controller in registry.values()]
+
+
+def safeupdate(*configs):
+    assert len(configs) > 1
+
+    def _merge(base, new):
+        assert isinstance(base, dict), f"{base} is not a dict"
+        assert isinstance(new, dict), f"{new} is not a dict"
+        out = deepcopy(base)
+        for k, v in new.items():
+            # assert k in out, f"{k} not in {base}"
+            if isinstance(v, dict):
+                if "grid_search" in v:
+                    out[k] = v
+                else:
+                    out[k] = _merge(out[k], v)
+            else:
+                out[k] = v
+
+        return out
+
+    return reduce(_merge, configs)
